Create Combat World Subsystem
The connections are still a bit rough, but we want the subsystem to manage the combat flow and calculations, broadcasting events when appropriate.
It is seeming useful to enable pawn manipulation throught controller states. Direct adjustments are happening currently, if there are issues we can setup helper funtions on the player controller